server/requirements.txt
fastapi
uvicorn[standard]
pydantic

python-jose[cryptography]

python-dotenv

httpx
pytest
pytest-asyncio